Refactor stats services
